Refractor and cleanup using common helper funcs. Tests passed
